+++ reqwest-retry/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2024-06-28
+
 ### Added
 - Added `with_retry_log_level` to `RetryTransientMiddleware`